Refactored multimedia into voice.
