+++ bundles/io/org.eclipse.smarthome.io.console/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Import-Package: org.apache.commons.lang,
  org.eclipse.smarthome.core.types,
  org.eclipse.smarthome.io.console,
  org.eclipse.smarthome.io.console.extensions,
- org.eclipse.smarthome.io.multimedia.text,
  org.osgi.framework,
  org.osgi.util.tracker,
  org.slf4j
