+++ bundles/io/org.eclipse.smarthome.io.voice/.classpath
