+++ bundles/io/org.eclipse.smarthome.io.voice/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.eclipse.smarthome.io.multimedia</name>
+	<name>org.eclipse.smarthome.io.voice</name>
 	<comment></comment>
 	<projects>
 	</projects>
