+++ bundles/io/org.eclipse.smarthome.io.voice/.settings/org.eclipse.jdt.core.prefs
