+++ bundles/io/org.eclipse.smarthome.io.voice/.settings/org.eclipse.pde.core.prefs
