+++ bundles/io/org.eclipse.smarthome.io.voice/META-INF/MANIFEST.MF
@@ -1,7 +1,7 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: Eclipse SmartHome Multimedia I/O Bundle
-Bundle-SymbolicName: org.eclipse.smarthome.io.multimedia
+Bundle-Name: Eclipse SmartHome Voice I/O Bundle
+Bundle-SymbolicName: org.eclipse.smarthome.io.voice
 Bundle-Version: 0.8.0.qualifier
 Bundle-Vendor: Eclipse.org/SmartHome
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
@@ -15,12 +15,13 @@ Import-Package: org.apache.commons.collections,
  org.eclipse.smarthome.core.items.events,
  org.eclipse.smarthome.core.library.types,
  org.eclipse.smarthome.core.types,
- org.eclipse.smarthome.io.multimedia.text,
- org.eclipse.smarthome.io.multimedia.tts,
+ org.eclipse.smarthome.io.console,
+ org.eclipse.smarthome.io.console.extensions,
+ org.eclipse.smarthome.io.voice.tts,
  org.osgi.framework,
  org.slf4j
-Export-Package: org.eclipse.smarthome.io.multimedia.text,
- org.eclipse.smarthome.io.multimedia.tts
+Export-Package: org.eclipse.smarthome.io.voice.text,
+ org.eclipse.smarthome.io.voice.tts
 Service-Component: OSGI-INF/*.xml
 Bundle-ActivationPolicy: lazy
-Bundle-Activator: org.eclipse.smarthome.io.multimedia.internal.MultimediaActivator
+Bundle-Activator: org.eclipse.smarthome.io.voice.internal.VoiceActivator
