+++ bundles/io/org.eclipse.smarthome.io.voice/OSGI-INF/InterpretConsoleCommandExtension.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2014-2015 openHAB UG (haftungsbeschraenkt) and others.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.voice.internal.extensions.interpret">
+   <implementation class="org.eclipse.smarthome.io.voice.internal.extensions.InterpretConsoleCommandExtension"/>
+   <service>
+      <provide interface="org.eclipse.smarthome.io.console.extensions.ConsoleCommandExtension"/>
+   </service>
+</scr:component>
