+++ bundles/io/org.eclipse.smarthome.io.voice/OSGI-INF/SayConsoleCommandExtension.xml
@@ -8,11 +8,10 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.multimedia.internal.text.StandardHumanLanguageInterpretereter">
-   <implementation class="org.eclipse.smarthome.io.multimedia.internal.text.StandardHumanLangaugeInterpreter"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.voice.extensions.say">
+   <implementation class="org.eclipse.smarthome.io.voice.internal.extensions.SayConsoleCommandExtension"/>
    <service>
-      <provide interface="org.eclipse.smarthome.io.multimedia.text.HumanLanguageInterpreter"/>
+      <provide interface="org.eclipse.smarthome.io.console.extensions.ConsoleCommandExtension"/>
    </service>
    <reference bind="setItemRegistry" cardinality="1..1" interface="org.eclipse.smarthome.core.items.ItemRegistry" name="ItemRegistry" policy="static" unbind="unsetItemRegistry"/>
-   <reference bind="setEventPublisher" cardinality="1..1" interface="org.eclipse.smarthome.core.events.EventPublisher" name="EventPublisher" policy="static" unbind="unsetEventPublisher"/>
 </scr:component>
