+++ bundles/io/org.eclipse.smarthome.io.voice/OSGI-INF/StandardHumanLanguageInterpreter.xml
@@ -8,10 +8,11 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.console.internal.extension.InterpretConsoleCommandExtension">
-   <implementation class="org.eclipse.smarthome.io.console.internal.extension.InterpretConsoleCommandExtension"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.voice.text.stdinterpreter">
+   <implementation class="org.eclipse.smarthome.io.voice.internal.text.StandardHumanLanguageInterpreter"/>
    <service>
-      <provide interface="org.eclipse.smarthome.io.console.extensions.ConsoleCommandExtension"/>
+      <provide interface="org.eclipse.smarthome.io.voice.text.HumanLanguageInterpreter"/>
    </service>
    <reference bind="setItemRegistry" cardinality="1..1" interface="org.eclipse.smarthome.core.items.ItemRegistry" name="ItemRegistry" policy="static" unbind="unsetItemRegistry"/>
+   <reference bind="setEventPublisher" cardinality="1..1" interface="org.eclipse.smarthome.core.events.EventPublisher" name="EventPublisher" policy="static" unbind="unsetEventPublisher"/>
 </scr:component>
