+++ bundles/io/org.eclipse.smarthome.io.voice/build.properties
