+++ bundles/io/org.eclipse.smarthome.io.voice/src/main/java/org/eclipse/smarthome/io/voice/text/InterpretationException.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.io.multimedia.text;
+package org.eclipse.smarthome.io.voice.text;
 
 /**
  * An exception used by {@link HumanLanguageInterpreter}s, if an error occurs.
